RATEST-294:Update Login.feature file with Login functionalities
