+++ qaframework-bdd-tests/src/test/java/org/openmrs/contrib/qaframework/automation/LoginSteps.java
@@ -10,6 +10,7 @@
 package org.openmrs.contrib.qaframework.automation;
 
 import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
 
 import io.cucumber.java.After;
 import io.cucumber.java.en.And;
@@ -18,37 +19,41 @@ import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
 
 import org.openmrs.contrib.qaframework.RunTest;
+import org.openmrs.contrib.qaframework.page.HomePage;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 
 public class LoginSteps extends Steps {
 
+	private HomePage homePage;
+	
 	@After(RunTest.HOOK.SELENIUM)
 	public void destroy() {
 		quit();
 	}
-
+	
 	private void enterUsername(String username) {
 		driver.findElement(By.id("username")).sendKeys(username);
 	}
-
+	
 	private void enterPassword(String password) {
 		driver.findElement(By.id("password")).sendKeys(password);
 	}
-
+	
 	private WebElement getLoginButton() {
 		return getElement(By.id("loginButton"));
 	}
-
+	
 	private WebElement getLogOutLink() {
 		return getElement(By.className("logout"));
 	}
 
+	
 	@Given("User visits login page")
 	public void visitLoginPage() throws Exception {
 		goToLoginPage();
 	}
-
+	
 	@When("User enters {string} username")
 	public void anyUsername(String username) {
 		if ("setupUser".equals(username)) {
@@ -56,7 +61,7 @@ public class LoginSteps extends Steps {
 		}
 		enterUsername(username);
 	}
-
+	
 	@And("User enters {string} password")
 	public void anyPassword(String password) {
 		if ("setupPass".equals(password)) {
@@ -64,7 +69,7 @@ public class LoginSteps extends Steps {
 		}
 		enterPassword(password);
 	}
-
+	
 	@And("User Selects {string} Login Location")
 	public void selectLoginLocation(String loginLocation) {
 		if ("firstLocation".equals(loginLocation)) {
@@ -78,12 +83,18 @@ public class LoginSteps extends Steps {
 			loginPage.clickOn(By.id(loginLocation));
 		}
 	}
-
+	
 	@And("User Logs in")
 	public void userLogsIn() {
 		getLoginButton().click();
 	}
 
+	public void initiateHomePage(){
+		homePage = new HomePage(page);
+		assertPage(homePage.waitForPage());
+
+	}
+	
 	@Then("System Evaluates Login {string}")
 	public void evaluateLogin(String status) {
 		if (status.trim().endsWith("true")) {
@@ -91,5 +102,41 @@ public class LoginSteps extends Steps {
 		} else if (status.trim().endsWith("false")) {
 			assertNotNull(getLoginButton());
 		}
+	}
+	
+	@Then("User logs in as clerk")
+	public void verifyClerkModulesAvailableOnHomePage() {
+		goToLoginPage().loginAsClerk();
+		initiateHomePage();
+		assertTrue(homePage.isActiveVisitsAppPresent());
+		assertTrue(homePage.isAppointmentSchedulingAppPresent());
+		assertTrue(homePage.isRegisterPatientCustomizedForRefAppPresent());
+	}
+	
+	@And("User logs in as doctor")
+	public void verifyDoctorModulesAvailableOnHomePage() {
+		goToLoginPage().loginAsDoctor();
+		initiateHomePage();
+		assertTrue(homePage.isFindAPatientAppPresent());
+		assertTrue(homePage.isActiveVisitsAppPresent());
+		assertTrue(homePage.isAppointmentSchedulingAppPresent());
+	}
+	
+	@And("User logs in as Nurse")
+	public void verifyNurseModulesAvailableOnHomePage() {
+		goToLoginPage().loginAsNurse();
+		initiateHomePage();
+		assertTrue(homePage.isFindAPatientAppPresent());
+		assertTrue(homePage.isActiveVisitsAppPresent());
+		assertTrue(homePage.isAppointmentSchedulingAppPresent());
+		assertTrue(homePage.isCaptureVitalsAppPresent());
+	}
+	
+	@Then("User logs in system admin")
+	public void verifySysadminModulesAvailableOnHomePage() {
+		goToLoginPage().loginAsSysadmin();
+		initiateHomePage();
+		assertTrue(homePage.isAppointmentSchedulingAppPresent());
+		assertTrue(homePage.isSystemAdministrationAppPresent());
 	}
 }
